pyproject.toml
[tool.poetry]
name = "rdc-video-bot"
version = "0.1.0"
description = ""
authors = ["Shargrove09 <8881163+Shargrove09@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
google-auth-oauthlib = "^1.0.0"
google-api-python-client = "^2.0.0"
pandas = "^2.0.0"
python-dotenv = "^1.0.0"
gspread = "^5.0.0"
gspread-dataframe = "^3.0.0"
rapidfuzz = "^3.0.0"
colorama = "^0.4.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
